\name{MoveFront}
\alias{MoveFront}
\title{Move a variable to the front of a data frame.}
\usage{
  MoveFront(data, Var)
}
\arguments{
  \item{data}{a data frame object}

  \item{Var}{the variable you would like to move to the
  front of the data frame.}
}
\description{
  \code{MoveFront} moves a specified variable to the front
  of a data frame.
}

