% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull}
\alias{pull}
\title{Extract a single column from a data frame (including tbl_df) and return it as
a vector.}
\source{
Modified from Tommy O'Dell:
\url{http://stackoverflow.com/a/24730843/1705044}
}
\usage{
pull(data, Var)
}
\arguments{
\item{data}{a data frame to extract the column from.}

\item{Var}{a character string identifying the column to extract as a vector.}
}
\value{
a vector
}
\description{
Extract a single column from a data frame (including tbl_df) and return it as
a vector.
}

