% Generated by roxygen2 (4.0.0): do not edit by hand
\name{shiftMA}
\alias{shiftMA}
\title{Internal function for slideMA}
\usage{
shiftMA(x, shiftBy, Abs, reminder)
}
\arguments{
\item{x}{vector}

\item{shiftBy}{numeric}

\item{Abs}{numeric}

\item{reminder}{logical}
}
\description{
Internal function for slideMA
}
\keyword{internals}

