% Generated by roxygen2 (4.0.2): do not edit by hand
\name{FillDown}
\alias{FillDown}
\title{Fills in missing (NA) values with the previous non-missing value}
\usage{
FillDown(data, Var)
}
\arguments{
\item{data}{a data frame.}

\item{Var}{the variable in \code{data} you would like to fill down missing
(\code{NA}) values.}
}
\value{
data frame
}
\description{
Fills in missing (NA) values with the previous non-missing value
}
\examples{
# Create fake data
id <- c('Algeria', NA, NA, NA, 'Mexico', NA, NA)
score <- rnorm(7)
Data <- data.frame(id, score)

# FillDown id
DataOut <- FillDown(Data, 'id')
}

